Min-Max normalization

+++ data/dataset.py
@@ -23,19 +23,19 @@ class MolDataSet(Dataset):
         nmol = len(molecules)
         ndesc = len(desc_mapping.desc_mapping)
         # bags: Nmol*Nconf*Ndesc 训练数据
-        self.bags = torch.from_numpy(np.zeros((nmol, nconf, ndesc),dtype=np.float32))
+        self.bags = torch.from_numpy(np.zeros((nmol, nconf, ndesc),dtype=np.uint8))
         # mask: Nmol*Nconf*1 标记哪些构象是有效的，在训练过程中去除噪点
-        self.mask = torch.from_numpy(np.ones((nmol, nconf,1),dtype=np.float32))
+        self.mask = torch.from_numpy(np.ones((nmol, nconf,1),dtype=np.uint8))
         # labels: Nmol
         self.labels = torch.from_numpy(np.zeros(nmol,dtype=np.float32))
         for i,molecule in enumerate(molecules):
             mol = molecule.mol
             self.labels[i] = float(molecule.activity)
-            self.mask[i][mol.GetNumConformers():] = 0.0
+            self.mask[i][mol.GetNumConformers():] = 0
             for conf in mol.GetConformers():
                 descs = molecule.get_conf_desc(conf.GetId())
                 for index,amount in descs.items():
-                    self.bags[i,int(conf.GetId()),int(index)] = float(amount)    
+                    self.bags[i,int(conf.GetId()),int(index)] = amount   
                    
     def __len__(self):
         return self.bags.shape[0]
